Add Notes and stuff, start building out songs

+++ Assets/Script/SongSO.cs
@@ -7,6 +7,6 @@ public class SongSO : ScriptableObject
     public AudioClip SongClip;
 
     public int BeatsPerMinute;
-    public List<decimal> NoteTimes;
+    public List<float> NoteTimes;
 
 }
